+++ app/src/main/java/com/bimapp/model/data_access/SyncAdapter.java
@@ -8,9 +8,6 @@ import android.content.ContentResolver;
 import android.content.Context;
 import android.content.SyncResult;
 import android.os.Bundle;
-import android.util.Log;
-
-import com.bimapp.BimApp;
 
 /**
  * Handle the transfer of data between a server and an
@@ -23,8 +20,6 @@ public class SyncAdapter extends AbstractThreadedSyncAdapter {
 
     private final AccountManager mAccountManager;
 
-    private BimApp mContext;
-
     /**
      * Set up the sync adapter
      */
@@ -36,8 +31,6 @@ public class SyncAdapter extends AbstractThreadedSyncAdapter {
          */
         mContentResolver = context.getContentResolver();
         mAccountManager = AccountManager.get(context);
-        mContext = (BimApp) context;
-        Log.d("SyncAdapter", "Created SyncAdapter") ;
     }
 
 
@@ -57,8 +50,6 @@ public class SyncAdapter extends AbstractThreadedSyncAdapter {
          */
         mContentResolver = context.getContentResolver();
         mAccountManager = AccountManager.get(context);
-        mContext = (BimApp) context;
-        Log.d("SyncAdapter", "Created SyncAdapter, Compat ");
     }
 
     @Override
@@ -67,21 +58,6 @@ public class SyncAdapter extends AbstractThreadedSyncAdapter {
                               String authority,
                               ContentProviderClient provider,
                               SyncResult syncResult) {
-
-        Log.d("SyncAdapter", "Started syncing");
-        // Should somehow determine if server and local db is different
-
-
-        // Get Projects/IssueBoards
-
-        // Get Topics
-
-        // Get Comments
-
-        // Get ViewPoints
-
-        // Determine what needs to be uploaded
-
 
 
     }
