+++ app/src/main/java/com/bimapp/model/data_access/entityManagers/ProjectEntityManager.java
@@ -5,7 +5,6 @@ import android.util.Log;
 import com.android.volley.Request;
 import com.bimapp.BimApp;
 import com.bimapp.controller.interfaces.ProjectsFragmentInterface;
-import com.bimapp.model.data_access.DataProvider;
 import com.bimapp.model.data_access.network.APICall;
 import com.bimapp.model.data_access.network.Callback;
 import com.bimapp.model.data_access.network.NetworkConnManager;
@@ -25,11 +24,8 @@ public class ProjectEntityManager implements ProjectsFragmentInterface.FragmentP
 
     private BimApp mContext;
 
-    private ProjectDBHandler handler;
-
     public ProjectEntityManager(BimApp context) {
         mContext = context;
-        handler = new ProjectDBHandler(context.getContentResolver());
     }
 
     @Override
@@ -73,8 +69,6 @@ public class ProjectEntityManager implements ProjectsFragmentInterface.FragmentP
             } catch (JSONException e) {
                 e.printStackTrace();
             }
-            for(Project p : projects)
-                handler.startInsert(1,mControllerCallback, DataProvider.ParseUri(DataProvider.PROJECT_TABLE),p.getContentValues() );
             mControllerCallback.setProjects(projects);
 
         }
@@ -86,8 +80,6 @@ public class ProjectEntityManager implements ProjectsFragmentInterface.FragmentP
      * @param controllerCallback is an interface that defines the methods for handling the responses from the {@Link NetworkConnManager}
      */
     public void getProjects(ProjectsFragmentInterface controllerCallback) {
-        handler.startQuery(1,controllerCallback, DataProvider.ParseUri(DataProvider.PROJECT_TABLE),
-                null,null,null,null);
         NetworkConnManager.networkRequest(mContext, Request.Method.GET, APICall.GETProjects(),
                 new ProjectCallback(controllerCallback), null);
     }
